erdrs: forder & crate rename
