+++ framework/snippets/src/interactor_retrieve.rs
@@ -1,6 +1,6 @@
 use crate::Interactor;
 use log::info;
-use mx_sdk_erdrs::data::transaction::TransactionOnNetwork;
+use mx_sdk::data::transaction::TransactionOnNetwork;
 use std::time::Duration;
 
 const TX_GET_RESULTS_NUM_RETRIES: usize = 8;
