+++ framework/snippets/src/interactor_vm_query.rs
@@ -7,7 +7,7 @@ use mx_sc_scenario::{
     },
     DebugApi,
 };
-use mx_sdk_erdrs::data::vm::VmValueRequest;
+use mx_sdk::data::vm::VmValueRequest;
 
 impl Interactor {
     pub async fn vm_query<CC, RequestedResult>(&mut self, contract_call: CC) -> RequestedResult
