+++ sdk/core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "mx-sdk-erdrs"
+name = "mx-sdk"
 version = "0.2.1"
 edition = "2021"
