+++ sdk/core/src/crypto/edwards25519/completed_group_element.rs
