+++ sdk/core/src/crypto/edwards25519/constant.rs
