+++ sdk/core/src/crypto/edwards25519/extended_group_element.rs
