+++ sdk/core/src/crypto/edwards25519/mod.rs
