+++ sdk/core/src/crypto/edwards25519/pre_computed_group_element.rs
