+++ sdk/core/src/data/esdt.rs
