+++ sdk/core/src/data/network_config.rs
