+++ sdk/core/src/data/network_status.rs
