+++ sdk/core/src/data/vm.rs
